Fixed startup issue

+++ src/NetworkInterface.cpp
@@ -1948,14 +1948,14 @@ int NetworkInterface::getFlows(lua_State* vm,
   lua_newtable(vm);
 
   if(a2zSortOrder) {
-    for(int i=toSkip, num=0; i<retriever.actNumEntries; i++) {
+    for(int i=toSkip, num=0; i<(int)retriever.actNumEntries; i++) {
       retriever.elems[i].flow->lua(vm, allowed_hosts, highDetails, true);
-      if(++num >= maxHits) break;
+      if(++num >= (int)maxHits) break;
     }
   } else {
     for(int i=(retriever.actNumEntries-1-toSkip), num=0; i>=0; i--) {
       retriever.elems[i].flow->lua(vm, allowed_hosts, highDetails, true);
-      if(++num >= maxHits) break;
+      if(++num >= (int)maxHits) break;
     }
   }
 
@@ -2010,16 +2010,16 @@ int NetworkInterface::getActiveHostsList(lua_State* vm, patricia_tree_t *allowed
   lua_newtable(vm);
 
   if(a2zSortOrder) {
-    for(int i=toSkip, num=0; i<retriever.actNumEntries; i++) {
+    for(int i=toSkip, num=0; i<(int)retriever.actNumEntries; i++) {
       retriever.elems[i].hostValue->lua(vm, NULL /* Already checked */, 
 					host_details, false, false, true);
-      if(++num >= maxHits) break;
+      if(++num >= (int)maxHits) break;
     }
   } else {
     for(int i=(retriever.actNumEntries-1-toSkip), num=0; i>=0; i--) {
       retriever.elems[i].hostValue->lua(vm, NULL /* Already checked */,
 					host_details, false, false, true);
-      if(++num >= maxHits) break;
+      if(++num >= (int)maxHits) break;
     }
   }
